[FEATURE] Update tstamp in record table
A new constant is introduced that enables the updating of the tstamp
field for the address records:
plugin.formhandler_subscription.updateDatabaseTimestamps

THe ValidateAuthCodeDB preprocessor now accepts a new config option
"updateTimestampOnRecordActivation". If this is TRUE the tstamp
field in the related record will be updated if configured in the TCA.

By default this constant is disabled to keep backwards compatibility.