[BUGFIX] Invalid direct_mail category select
The configuration of the select statement that loads the currently
selected direct_mail categories of the user from the database
in for the update form is fixed.

With the old configuration an error occured:
There is no entry in the $TCA array for the table
"sys_dmail_ttaddress_category_mm".

Thanks to Oliver Hübner for providing the patch!

Relates: #65664